Reuse doctrine's connection params
Currently the user of this bundle has to config the database connection params additionally to the doctrine connection if the `DATABASE_URL` is used instead of the specific params. This change adds an auto detection of the connection params used for doctrine. In the case the sulu database differs from the default connection the correct one can be specified in the bundle's config.